+++ src/render/prep_render.c
@@ -1,0 +1,51 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   prep_render.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/06/24 23:08:02 by reasuke           #+#    #+#             */
+/*   Updated: 2024/06/24 23:16:56 by reasuke          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include <math.h>
+
+#include "ctx.h"
+#include "libft.h"
+#include "mlx.h"
+
+#include "render_internal.h"
+
+static void	_reset_points(t_ctx *ctx)
+{
+	ft_bzero(ctx->mlx_conf->p_data,
+		WIN_WIDTH * WIN_HEIGHT * (ctx->mlx_conf->bits_per_pixel / 8));
+	clean_points(ctx->data);
+	ctx->view_conf->offset_x = 0;
+	ctx->view_conf->offset_y = 0;
+	init_points(ctx->data);
+	init_colors(ctx->data, ctx->view_conf->preset);
+}
+
+void	prep_render(t_ctx *ctx, bool is_clean)
+{
+	if (is_clean)
+		_reset_points(ctx);
+	if (ctx->view_conf->type == ISOMETRIC)
+	{
+		apply_operation(ctx->data, rotate_z, M_PI_4);
+		apply_operation(ctx->data, rotate_x, atan(1 / sqrt(2)));
+		render_image(ctx);
+		mlx_string_put(ctx->mlx_conf->p_mlx, ctx->mlx_conf->p_win,
+			920, 20, 0xFFFFFF, "isometric view");
+	}
+	else if (ctx->view_conf->type == TOPVIEW)
+	{
+		render_image(ctx);
+		mlx_string_put(ctx->mlx_conf->p_mlx, ctx->mlx_conf->p_win,
+			920, 20, 0xFFFFFF, "top view");
+	}
+	render_instructions(ctx);
+}
